Test code style example files on CI
