+++ lib/launch_base/version.rb
@@ -1,0 +1,3 @@
+module LaunchBase
+  VERSION = "0.1.0"
+end
