+++ package.json
@@ -1,0 +1,17 @@
+{
+  "name": "launch-base",
+  "version": "0.0.1",
+  "description": "Kabisa Launch Base",
+  "main": "index.js",
+  "repository": "git@github.com:kabisa/launch-base.git",
+  "author": "Martijn Versluis <martijnversluis@users.noreply.github.com>",
+  "license": "MIT",
+  "devDependencies": {
+    "babel-eslint": "^8.2.3",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0"
+  }
+}
