+++ templates/linters/.codeclimate.yml
