+++ templates/linters/.eslintrc.json
