+++ templates/linters/.rubocop.yml
