+++ test/linters/es6.js
@@ -1,0 +1,5 @@
+function someFunction() {
+  return 'This is a very long line in order to test whether or not eslint accepts 120 characters on a single line.....';
+}
+
+export default someFunction();
